Refactor all code for retrieval => move logic in server side

+++ grails-app/domain/be/cytomine/ontology/Term.groovy
@@ -13,7 +13,7 @@ class Term extends SequenceDomain implements Serializable {
     Ontology ontology
     String color
 
-    Integer rate
+    Double rate
 
     static belongsTo = [ontology:Ontology]
     static transients = ["rate"]
